selenium/requirements.txt
selenium==4.10.0
python-dotenv==1.0.0